backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "girisushil",email = "girisushil23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, !=3.14.1, <3.15"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "elasticsearch (>=9.1.1,<10.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "antlr4-python3-runtime (>=4.13.2,<5.0.0)",
    "ruamel-yaml (>=0.18.16,<0.19.0)",
    "requests (>=2.32.5,<3.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "checkov (>=3.2.495,<4.0.0)",
    "litellm (>=1.80.5,<2.0.0)",
    "python-hcl2 (>=7.3.1,<8.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tqdm (>=4.67.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
